Adding Rake as a depencency in Gemfile, updating Docs
If using bundler for dependency management, you also need
to call the bundler versions of the binaries.

Installing rake as a bundler dep gives us the possibility
of calling "bundle exec rake" which will then, automatically,
use bundled binaries if called inside the rake file (when called
with system() or such).

This should make the 3 step setup instruction from the README
finally valid.